Hides debug heatmap if not verbose; adds option to hide freq_margins; heatmap formatting.
Note: sets vmin to 0 for heatmaps, including with number of pairs. The previous plot was treacherous as it used the full gradient for whatever variation we had: light colors looked like we had little data, when it could still mean thousands of pairs. This will be more interpretable at a glance (although it hides fine variation in number of pairs)